refactor(frontend): move the allergy panel's state into a hook
AllergyListPanel combined permission checks, loading, fetching, creation,
resolution, notifications and rendering in one 114-line component. Move the
state into useAllergyList so the panel is a thin projection onto the
presentational list, matching the waitlist panel's shape.

Behaviour is unchanged - the panel's own tests pass untouched (318 tests across
the companion-history suites), with the panel at 100% and the hook at 100%
statements, functions and lines. tsc 0, lint clean.
